requirements.txt
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.4
filelock==3.15.1
fsspec==2024.6.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
gTTS==2.5.1
httplib2==0.22.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.24.3
oauthlib==3.2.2
openai-whisper==20231117
playsound==1.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pyparsing==3.1.2
pypiwin32==223
pyttsx3==2.90
pytz==2024.1
pywin32==306
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
soundfile==0.12.1
SpeechRecognition==3.10.4
sympy==1.12.1
tbb==2021.12.0
tiktoken==0.7.0
torch==2.3.1
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
